The Elara OS kernel now has ext4 support to load the non firmware kernel from disk.

+++ libElaraParallelAssembly/src/opcodes/epa_opcode_values.h
@@ -4,10 +4,10 @@
 //
 // Address space policy unchanged
 //
-#define EPA_OP_RESERVED_MIN ((uint8_t)128u)
-#define EPA_OP_EXT_ESCAPE   ((uint8_t)255u)
+#define EPA_OP_RESERVED_MIN ((uint16_t)128u)
+#define EPA_OP_EXT_ESCAPE   ((uint16_t)255u)
 
-#define EPA_OP_IS_COMMON(op) ((uint8_t)(op) < EPA_OP_RESERVED_MIN)
+#define EPA_OP_IS_COMMON(op) ((uint16_t)(op) < EPA_OP_RESERVED_MIN)
 #define EPA_OP_IS_GL(op)     (0)
 #define EPA_OP_IS_CUDA(op)   (0)
 
@@ -20,7 +20,7 @@
 // Opcode numeric values
 //
 #define X(name, value, mnemonic, plen) \
-  enum { EPA_OP_##name = (uint8_t)(value) };
+  enum { EPA_OP_##name = (uint16_t)(value) };
 
 #include "epa_opcodes.x"
